+++ packages/flutter_meedu/example/lib/pages/home/home_provider.dart
@@ -1,7 +1,10 @@
 import 'package:meedu/meedu.dart';
 import 'package:meedu_example/pages/home/home_controller.dart';
 import 'package:meedu_example/pages/home/home_state.dart';
+import 'package:meedu_example/pages/login/login_provider.dart';
 
 final homeProvider = StateProvider<HomeController, HomeState>(
-  (_) => HomeController(),
+  (ref) => HomeController(
+    ref.read(loginProvider),
+  ),
 );