+++ packages/meedu/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [3.0.0-prev.1]
+- Fixed `read` method into `ProviderReference`.
+
 ## [3.0.0-prev.0]
 - Prev version of 3.0.0